+++ lib/lldb/error.rb
@@ -13,6 +13,36 @@ module LLDB
   class UnsupportedPlatformError < LLDBError; end
   class IncompatibleWrapperError < LLDBError; end
 
+  class OperationError < LLDBError
+    attr_reader :operation, :error, :code, :error_type
+
+    # @rbs operation: String
+    # @rbs error: Error
+    # @rbs return: void
+    def initialize(operation, error)
+      @operation = operation
+      @error = error
+      @code = error.code
+      @error_type = error.type
+      super("LLDB operation '#{operation}' failed (type=#{error.type_name}, code=#{error.code}): #{error.message}")
+    end
+  end
+
+  class InternalBindingError < LLDBError
+    attr_reader :operation, :native_message, :native_code
+
+    # @rbs operation: String
+    # @rbs native_message: String
+    # @rbs native_code: Integer
+    # @rbs return: void
+    def initialize(operation, native_message, native_code)
+      @operation = operation
+      @native_message = native_message
+      @native_code = native_code
+      super("Native wrapper operation '#{operation}' failed (code=#{native_code}): #{native_message}")
+    end
+  end
+
   class Error
     # @rbs ptr: FFI::Pointer?
     # @rbs return: void
@@ -45,8 +75,20 @@ module LLDB
     alias to_s message
 
     # @rbs return: Integer
-    def error_code
+    def code
       FFIBindings.lldb_error_get_error(@ptr)
+    end
+
+    alias error_code code
+
+    # @rbs return: Integer
+    def type
+      FFIBindings.lldb_error_get_type(@ptr)
+    end
+
+    # @rbs return: String
+    def type_name
+      ErrorType::NAMES.fetch(type, 'unknown')
     end
 
     # @rbs return: void
@@ -65,9 +107,10 @@ module LLDB
       @ptr
     end
 
+    # @rbs operation: String
     # @rbs return: void
-    def raise_if_error!
-      raise LLDBError, message if fail?
+    def raise_if_error!(operation = 'native operation')
+      raise OperationError.new(operation, self) if fail?
     end
   end
 end